requirements.txt
annotated-types==0.7.0
anyio==4.7.0
click==8.1.8
colorama==0.4.6
databases==0.9.0
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.6
greenlet==3.1.1
h11==0.14.0
idna==3.10
mysql-connector-python==9.1.0
pydantic==2.10.4
pydantic_core==2.27.2
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.34.0
langchain_community
transformers
langchain_ollama
